requirements.txt
keras==2.12.0
numpy==1.23.5
Pillow==9.5.0
Pillow==9.5.0
streamlit==1.22.0
